Since we changed 'createTime' property from milliseconds to seconds, we can downscale the type from 'long' to 'uint'.

+++ WonkaSystem/WonkaEth/Contracts/WonkaRuleTreeRegistry.cs
@@ -52,7 +52,7 @@ namespace WonkaEth.Contracts
 
         public HashSet<string> UsedCustomOps { get; set; }
 
-        public long creationTime;
+        public uint creationTime;
     }
 
     public class WonkaRuleTreeRegistry
